Refactor platform handling: replace Platform with MobileSubscriptionPlatform and update related tests
