--- membership-attribute-service/test/controllers/AttributeControllerLogicTest.scala
@@ -1,2 +1,0 @@
-package controllers
-
